requirements.txt
crewai
mysql-connector-python
